getinfo {lightgbm}	R Documentation

Get information of an lgb.Dataset object

Description

Get information of an lgb.Dataset object

Usage

getinfo(dataset, ...)

## S3 method for class 'lgb.Dataset'
getinfo(dataset, name, ...)

Arguments

dataset
Object of class lgb.Dataset

...
other parameters

name
the name of the information field to get (see details)

Details

The name field can be one of the following:

label: label lightgbm learn from ;

weight: to do a weight rescale ;

group: group size

init_score: initial score is the base prediction lightgbm will boost from ;

Value

info data

Examples

## Not run: 
library(lightgbm)
data(agaricus.train, package = "lightgbm")
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label = train$label)
lgb.Dataset.construct(dtrain)

labels <- lightgbm::getinfo(dtrain, "label")
lightgbm::setinfo(dtrain, "label", 1 - labels)

labels2 <- lightgbm::getinfo(dtrain, "label")
stopifnot(all(labels2 == 1 - labels))

## End(Not run)

[Package lightgbm version 2.1.2 Index]