lgb.cv {lightgbm}	R Documentation

Main CV logic for LightGBM

Description

Simple interface for training an lightgbm model. Its documentation is combined with lgb.train.

Usage

lgb.cv(params = list(), data, nrounds = 10, nfold = 3, label = NULL,
  weight = NULL, obj = NULL, eval = NULL, verbose = 1, record = TRUE,
  eval_freq = 1L, showsd = TRUE, stratified = TRUE, folds = NULL,
  init_model = NULL, colnames = NULL, categorical_feature = NULL,
  early_stopping_rounds = NULL, callbacks = list(), ...)

lgb.train(params = list(), data, nrounds = 10, valids = list(),
  obj = NULL, eval = NULL, verbose = 1, record = TRUE, eval_freq = 1L,
  init_model = NULL, colnames = NULL, categorical_feature = NULL,
  early_stopping_rounds = NULL, callbacks = list(), reset_data = FALSE,
  ...)

lightgbm(data, label = NULL, weight = NULL, params = list(),
  nrounds = 10, verbose = 1, eval_freq = 1L,
  early_stopping_rounds = NULL, save_name = "lightgbm.model",
  init_model = NULL, callbacks = list(), ...)

Arguments

params
List of parameters

data
a lgb.Dataset object, used for CV

nrounds
number of CV rounds

nfold
the original dataset is randomly partitioned into nfold equal size subsamples.

label
vector of response values. Should be provided only when data is an R-matrix.

weight
vector of response values. If not NULL, will set to dataset

obj
objective function, can be character or custom objective function. Examples include regression, regression_l1, huber, binary, lambdarank, multiclass, multiclass

eval
evaluation function, can be (list of) character or custom eval function

verbose
verbosity for output, if <= 0, also will disable the print of evalutaion during training

record
Boolean, TRUE will record iteration message to booster$record_evals

eval_freq
evalutaion output frequence, only effect when verbose > 0

showsd
boolean, whether to show standard deviation of cross validation

stratified
a boolean indicating whether sampling of folds should be stratified by the values of outcome labels.

folds
list provides a possibility to use a list of pre-defined CV folds (each element must be a vector of test fold's indices). When folds are supplied, the nfold and stratified parameters are ignored.

init_model
path of model file of lgb.Booster object, will continue train from this model

colnames
feature names, if not null, will use this to overwrite the names in dataset

categorical_feature
list of str or int type int represents index, type str represents feature names

early_stopping_rounds
int Activates early stopping. Requires at least one validation data and one metric If there's more than one, will check all of them except the training data Returns the model with (best_iter + early_stopping_rounds) If early stopping occurs, the model will have 'best_iter' field

callbacks
list of callback functions List of callback functions that are applied at each iteration.

...
other parameters, see Parameters.rst for more informations

valids
a list of lgb.Dataset objects, used for validation

reset_data
Boolean, setting it to TRUE (not the default value) will transform the booster model into a predictor model which frees up memory and the original datasets

boosting
boosting type. gbdt, dart

num_leaves
number of leaves in one tree. defaults to 127

max_depth
Limit the max depth for tree model. This is used to deal with overfit when #data is small. Tree still grow by leaf-wise.

num_threads
Number of threads for LightGBM. For the best speed, set this to the number of real CPU cores, not the number of threads (most CPU using hyper-threading to generate 2 threads per CPU core).

params
List of parameters

data
a lgb.Dataset object, used for training

nrounds
number of training rounds

obj
objective function, can be character or custom objective function. Examples include regression, regression_l1, huber, binary, lambdarank, multiclass, multiclass

boosting
boosting type. gbdt, dart

num_leaves
number of leaves in one tree. defaults to 127

max_depth
Limit the max depth for tree model. This is used to deal with overfit when #data is small. Tree still grow by leaf-wise.

num_threads
Number of threads for LightGBM. For the best speed, set this to the number of real CPU cores, not the number of threads (most CPU using hyper-threading to generate 2 threads per CPU core).

eval
evaluation function, can be (a list of) character or custom eval function

verbose
verbosity for output, if <= 0, also will disable the print of evalutaion during training

record
Boolean, TRUE will record iteration message to booster$record_evals

eval_freq
evalutaion output frequency, only effect when verbose > 0

init_model
path of model file of lgb.Booster object, will continue training from this model

colnames
feature names, if not null, will use this to overwrite the names in dataset

categorical_feature
list of str or int type int represents index, type str represents feature names

early_stopping_rounds
int Activates early stopping. Requires at least one validation data and one metric If there's more than one, will check all of them except the training data Returns the model with (best_iter + early_stopping_rounds) If early stopping occurs, the model will have 'best_iter' field

callbacks
list of callback functions List of callback functions that are applied at each iteration.

...
other parameters, see Parameters.rst for more informations

Value

a trained model lgb.CVBooster.

a trained booster model lgb.Booster.

Examples

## Not run: 
library(lightgbm)
data(agaricus.train, package = "lightgbm")
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label = train$label)
params <- list(objective = "regression", metric = "l2")
model <- lgb.cv(params,
                dtrain,
                10,
                nfold = 5,
                min_data = 1,
                learning_rate = 1,
                early_stopping_rounds = 10)

## End(Not run)
## Not run: 
library(lightgbm)
data(agaricus.train, package = "lightgbm")
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label = train$label)
data(agaricus.test, package = "lightgbm")
test <- agaricus.test
dtest <- lgb.Dataset.create.valid(dtrain, test$data, label = test$label)
params <- list(objective = "regression", metric = "l2")
valids <- list(test = dtest)
model <- lgb.train(params,
                   dtrain,
                   100,
                   valids,
                   min_data = 1,
                   learning_rate = 1,
                   early_stopping_rounds = 10)

## End(Not run)

[Package lightgbm version 2.1.2 Index]