setinfo {lightgbm}	R Documentation

Set information of an lgb.Dataset object

Description

Set information of an lgb.Dataset object

Usage

setinfo(dataset, ...)

## S3 method for class 'lgb.Dataset'
setinfo(dataset, name, info, ...)

Arguments

dataset
Object of class "lgb.Dataset"

...
other parameters

name
the name of the field to get

info
the specific field of information to set

Details

The name field can be one of the following:

label: label lightgbm learn from ;

weight: to do a weight rescale ;

init_score: initial score is the base prediction lightgbm will boost from ;

group.

Value

passed object

Examples

## Not run: 
library(lightgbm)
data(agaricus.train, package = "lightgbm")
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label = train$label)
lgb.Dataset.construct(dtrain)

labels <- lightgbm::getinfo(dtrain, "label")
lightgbm::setinfo(dtrain, "label", 1 - labels)

labels2 <- lightgbm::getinfo(dtrain, "label")
stopifnot(all.equal(labels2, 1 - labels))

## End(Not run)

[Package lightgbm version 2.1.2 Index]