Bank Marketing Data Set

This data set is originally from the Bank Marketing data set, UCI Machine Learning Repository.

data(bank)

Format

A data.table with 4521 rows and 17 variables

Details

It contains only the following: bank.csv with 10 randomly selected from 3 (older version of this dataset with less inputs).

References

http://archive.ics.uci.edu/ml/datasets/Bank+Marketing

S. Moro, P. Cortez and P. Rita. (2014) A Data-Driven Approach to Predict the Success of Bank Telemarketing. Decision Support Systems