Predict method for LightGBM model

Predicted values based on class lgb.Booster

# S3 method for lgb.Booster
predict(object, data, num_iteration = NULL,
  rawscore = FALSE, predleaf = FALSE, header = FALSE, reshape = FALSE,
  ...)

Arguments

object
Object of class lgb.Booster

data
a matrix object, a dgCMatrix object or a character representing a filename

num_iteration
number of iteration want to predict with, NULL or <= 0 means use best iteration

rawscore
whether the prediction should be returned in the for of original untransformed sum of predictions from boosting iterations' results. E.g., setting rawscore=TRUE for logistic regression would result in predictions for log-odds instead of probabilities.

predleaf
whether predict leaf index instead.

header
only used for prediction for text file. True if text file has header

reshape
whether to reshape the vector of predictions to a matrix form when there are several prediction outputs per case.

Value

For regression or binary classification, it returns a vector of length nrows(data). For multiclass classification, either a num_class * nrows(data) vector or a (nrows(data), num_class) dimension matrix is returned, depending on the reshape value.

When predleaf = TRUE, the output is a matrix object with the number of columns corresponding to the number of trees.

Examples

# It crashes the documentation generation, skipping
if (!(pkgdown::in_pkgdown())) {

library(lightgbm)
data(agaricus.train, package = "lightgbm")
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label = train$label)
data(agaricus.test, package = "lightgbm")
test <- agaricus.test
dtest <- lgb.Dataset.create.valid(dtrain, test$data, label = test$label)
params <- list(objective = "regression", metric = "l2")
valids <- list(test = dtest)
model <- lgb.train(params,
                   dtrain,
                   100,
                   valids,
                   min_data = 1,
                   learning_rate = 1,
                   early_stopping_rounds = 10)
preds <- predict(model, test$data)
}